+++ conda/plugins/solvers/__init__.py
@@ -2,16 +2,16 @@
 # SPDX-License-Identifier: BSD-3-Clause
 """Register the classic conda solver."""
 
-from ..base.constants import CLASSIC_SOLVER
-from . import CondaSolver, hookimpl
+from ...base.constants import CLASSIC_SOLVER
+from .. import CondaSolver, hookimpl
 
 
 @hookimpl(tryfirst=True)  # make sure the classic solver can't be overwritten
 def conda_solvers():
     """The classic solver as shipped by default in conda."""
-    from ..core.solve import Solver
+    from .classic import ClassicSolver
 
     yield CondaSolver(
         name=CLASSIC_SOLVER,
-        backend=Solver,
+        backend=ClassicSolver,
     )
