+++ conda/plugins/solvers/classic/logic.py
